requirements.txt
python-telegram-bot>=22,<23
Pillow>=10,<12
dropbox>=12,<13
PyYAML>=6,<7
python-dotenv>=1,<2

